requirements.txt
streamlit==1.32.0
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
langdetect==1.0.9
tqdm==4.66.1
googletrans==3.1.0a0
python-dotenv==1.0.1
requests==2.31.0
plotly==5.18.0
